+++ contentcuration/contentcuration/static/js/edit_channel/channel_set/views/ChannelSetList.vue
@@ -30,7 +30,7 @@
     <hr/>
 
     <!-- Channel list section -->
-    <div class="channel-list">
+    <div class="channelset-list">
       <div v-if="loadChannels" class="default-item">
         {{ $tr('loading') }}
       </div>
