+++ contentcuration/contentcuration/static/less/channel_list.less
@@ -82,16 +82,16 @@ body {
 .channel-list {
     .channel-list-width;
     .new-button {
-	    text-align: right;
-	    margin-bottom: 30px;
-	    a, button {
-	      font-size: 14pt;
-	      span {
-	        font-size: 18pt;
-	      }
-	    }
-	  }
-	 .default-item {
+		text-align: right;
+		margin-bottom: 30px;
+		a, button {
+			font-size: 14pt;
+			span {
+				font-size: 18pt;
+			}
+		}
+	}
+	.default-item {
 		color: @gray-500;
 		font-size: 16pt;
 		font-style: italic;
